requirements.txt
lxml
requests
autoscraper
cssselect
uvicorn[standard]
fastapi
python-slugify
